exit advance, executable enleve

+++ src/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: tpaufert <tpaufert@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/24 16:14:17 by tpaufert          #+#    #+#             */
-/*   Updated: 2023/06/14 17:04:59 by tpaufert         ###   ########.fr       */
+/*   Updated: 2023/06/16 18:36:08 by tpaufert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,40 +22,40 @@
 #  define BUFFER_SIZE 50
 # endif
 
-int		ft_atoi(const char *str);
-void	ft_bzero(void *s, size_t n);
-int		ft_isalnum(int c);
-int		ft_isalpha(int c);
-int		ft_isascii(int c);
-int		ft_isdigit(int c);
-int		ft_isprint(int c);
-void	*ft_memchr(const void *s, int c, size_t n);
-int		ft_memcmp(const void *s1, const void *s2, size_t n);
-void	*ft_memcpy(void *dst, const void *src, size_t n);
-void	*ft_memmove(void *dst, const void *src, size_t len);
-void	*ft_memset( void *b, int c, size_t len);
-char	*ft_strchr(const char *s, int c);
-size_t	ft_strlen(const char *s);
-int		ft_strncmp(const char *s1, const char *s2, size_t n);
-char	*ft_strrchr(const char *s, int c);
-char	*ft_tolower(char *c);
-int		ft_toupper(int c);
-char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
-size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
-char	*ft_strjoin(char const *s1, char const *s2);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
-char	*ft_strtrim(char const *s1, char const *set);
-void	ft_putchar_fd(char c, int fd);
-void	ft_putstr_fd(char *s, int fd);
-void	ft_putendl_fd(char *s, int fd);
-void	ft_putnbr_fd(int n, int fd);
-char	*ft_strdup3(const char *s1);
-void	*ft_calloc(size_t count, size_t size);
-char	*ft_itoa(int n);
-char	**ft_split(char const *s, char c);
-size_t	ft_strlcat(char *dst, const char *src, size_t dstsize);
-char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
-void	ft_striteri(char *s, void (*f)(unsigned int, char*));
+long long int	ft_atoi(const char *str);
+void			ft_bzero(void *s, size_t n);
+int				ft_isalnum(int c);
+int				ft_isalpha(int c);
+int				ft_isascii(int c);
+int				ft_isdigit(int c);
+int				ft_isprint(int c);
+void			*ft_memchr(const void *s, int c, size_t n);
+int				ft_memcmp(const void *s1, const void *s2, size_t n);
+void			*ft_memcpy(void *dst, const void *src, size_t n);
+void			*ft_memmove(void *dst, const void *src, size_t len);
+void			*ft_memset( void *b, int c, size_t len);
+char			*ft_strchr(const char *s, int c);
+size_t			ft_strlen(const char *s);
+int				ft_strncmp(const char *s1, const char *s2, size_t n);
+char			*ft_strrchr(const char *s, int c);
+char			*ft_tolower(char *c);
+int				ft_toupper(int c);
+char			*ft_strnstr(const char *haystack, const char *needle, size_t len);
+size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
+char			*ft_strjoin(char const *s1, char const *s2);
+char			*ft_substr(char const *s, unsigned int start, size_t len);
+char			*ft_strtrim(char const *s1, char const *set);
+void			ft_putchar_fd(char c, int fd);
+void			ft_putstr_fd(char *s, int fd);
+void			ft_putendl_fd(char *s, int fd);
+void			ft_putnbr_fd(int n, int fd);
+char			*ft_strdup3(const char *s1);
+void			*ft_calloc(size_t count, size_t size);
+char			*ft_itoa(int n);
+char			**ft_split(char const *s, char c);
+size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
+char			*ft_strmapi(char const *s, char (*f)(unsigned int, char));
+void			ft_striteri(char *s, void (*f)(unsigned int, char*));
 
 typedef struct s_list
 {
